# inverter-mqtt/mqtt-push.sh: optimize