On Raspberry Pi 3B executing the original code, with Influx disabled, takes:
```
real    0m57,753s
user    0m54,679s
sys     0m2,214s
```
With the modification the time is:
```
real    0m10,733s
user    0m8,338s
sys     0m0,759s
```
For the record, I run the code outside of containers.